# Improved detect silence