@emsi 2024-02-05
It does not seem to work properly.
When ran on YouTube video (~4h length) with:
`split_on_silence(audio_segment, min_silence_len=800, keep_silence=True))`

It returns the following ranges (just 4 segments):
`(0, 6812736, 6812736, 13615464, 13615464, 13635080, 13635080, 13677621)`


When running the same file with the same arguments (`min_silence_len=800, silence_thresh=-16`) in Audacity it finds lots and lots of silence (and I can confirm at glance that those findings are correct):
![image](https://github.com/jiaaro/pydub/assets/433383/45dba82e-eb2c-4cba-8d39-eb81d5bbc61b)