@lumip 2024-02-24
> It does not seem to work properly. When ran on YouTube video (~4h length) with: `split_on_silence(audio_segment, min_silence_len=800, keep_silence=True))`
> 
> It returns the following ranges (just 4 segments): `(0, 6812736, 6812736, 13615464, 13615464, 13635080, 13635080, 13677621)`
> 
> When running the same file with the same arguments (`min_silence_len=800, silence_thresh=-16`) in Audacity it finds lots and lots of silence (and I can confirm at glance that those findings are correct): ![image](https://private-user-images.githubusercontent.com/433383/302339735-45dba82e-eb2c-4cba-8d39-eb81d5bbc61b.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MDg3NjEzODAsIm5iZiI6MTcwODc2MTA4MCwicGF0aCI6Ii80MzMzODMvMzAyMzM5NzM1LTQ1ZGJhODJlLWViMmMtNGNiYS04ZDM5LWViODFkNWJiYzYxYi5wbmc_WC1BbXotQWxnb3JpdGhtPUFXUzQtSE1BQy1TSEEyNTYmWC1BbXotQ3JlZGVudGlhbD1BS0lBVkNPRFlMU0E1M1BRSzRaQSUyRjIwMjQwMjI0JTJGdXMtZWFzdC0xJTJGczMlMkZhd3M0X3JlcXVlc3QmWC1BbXotRGF0ZT0yMDI0MDIyNFQwNzUxMjBaJlgtQW16LUV4cGlyZXM9MzAwJlgtQW16LVNpZ25hdHVyZT05MjY5M2RiOTJlNDc5YTI3MDJmOGM5NjJiY2QzZjc5YWY0ZmVhYWI3NWViMDJlZTViMjhjYTg3MzYzNDViNzMzJlgtQW16LVNpZ25lZEhlYWRlcnM9aG9zdCZhY3Rvcl9pZD0wJmtleV9pZD0wJnJlcG9faWQ9MCJ9.9gKgiZYud6_aG4dfBWRlRSEWCumWatLnbP8IR_XEmYY)

Hey, sorry I saw your responses a bit late just now. Could you perhaps provide a link to the video in question so that I can have a look?